requirements.txt
langchain
langchain-community
chromadb
pypdf
python-dotenv
langchain-groq
streamlit
rank_bm25
flashrank
pyyaml
ensure
python-box
ragas
datasets
pandas
rapidfuzz
langchain-ollama
langchain-experimental